% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_extraction_mapping_wrap.R
\name{extraction_annotation_pos}
\alias{extraction_annotation_pos}
\title{Extract germline mutations annotated by the snpEff tool for a cancer type of your interest.}
\usage{
extraction_annotation_pos(mutation_file, cancer_type, cancer_barcode,
  output_dir)
}
\arguments{
\item{mutation_file}{File name of the mc3 somatic variants from TCGA.}

\item{cancer_type}{TCGA cancer type identifier, e.g. STAD, BRCA.}

\item{output_dir}{The directory you would like to have your output files in.}

\item{cancer_barcod}{A list of barcodes user wants to process, matching the last column of mutation file.}
}
\description{
This function generates two files as the final products. First, mutations which result in protein-consequence with
effect on proteins annotated; Second, mutations without protein-consequence.
}
\examples{
extraction_annotation_pos(mutation_file = "/data/ginny/tcga_pancan/important_files/mc3_info.tsv" ,
                                 cancer_type = "STAD",
                                 cancer_barcode = stad_barcode,
                                 output_dir = "/data/ginny/tcga_pancan/STAD_somatic/")
}
