% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline_extraction_mapping_wrap.R
\name{germline_extraction_annotation_pos}
\alias{germline_extraction_annotation_pos}
\title{Extract germline mutations annotated by the snpEff tool for a cancer type of your interest.}
\usage{
germline_extraction_annotation_pos(raw_snpeff_output_dir, cancer_type,
  cancer_barcode, sample_cn_id, chr_name = c(seq(1:22), "X", "Y"),
  quality_filter = 30, output_dir, extraction_dir)
}
\arguments{
\item{raw_snpeff_output_dir}{Directory where raw snpeff outputs are stored。}

\item{cancer_type}{TCGA cancer type identifier, e.g. STAD, BRCA.}

\item{cancer_barcode}{TCGA barcodes for this cancer type cohort.}

\item{sample_cn_id}{A barcode codebook provided by the packge.}

\item{chr_name}{Which chromosomes are of interest, default is all of them.}

\item{quality_filter}{The minimiu value of QUAL for selection of mutations, default is 30}

\item{output_dir}{The directory you would like to have your output files in.}

\item{extraction_dir}{The directory you would like to have your raw extraction files in.}
}
\description{
This function generates two files as the final products. First, mutations which result in protein-consequence with
effect on proteins annotated; Second, mutations without protein-consequence.
}
