% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline_extraction_mapping_wrap.R
\name{germline_piu_mapping}
\alias{germline_piu_mapping}
\title{Map germline pc to piu and bpiu, summarise bpiu and npc per gene}
\usage{
germline_piu_mapping(ptm_domain_filename, pc_data_name, npc_data_name,
  cancer_barcode, sample_cn_id, output_dir)
}
\arguments{
\item{ptm_domain_filename}{The file provided by the package where the piu information is recorded.}

\item{pc_data_name}{the file listing mutations that result in protein consequence. A file generated by function germline_extraction_annotation_pos}

\item{npc_data_name}{the file listing mutations that result in non-protein consequence. A file generated by function germline_extraction_annotation_pos}

\item{cancer_barcode}{TCGA barcodes for this cancer type cohort.}

\item{sample_cn_id}{A barcode codebook provided by the packge.}

\item{output_dir}{The directory you would like to have your output files in.}
}
\description{
This function generates three files as the final products. First, each piu mapped, second, bpiu summarised per gene, third, 
npc summarised per gene.
}
\examples{
germline_piu_mapping (ptm_domain_filename =  "/data/ginny/tcga_pancan/important_files/ptm_domain_combine_df.tsv",
                     pc_data_name  = 
                       "/data/ginny/tcga_pancan/germline_raw_process/STAD_snpeff_annotation/snpeff_variant_anno_pc_pos.tsv",
                     npc_data_name = 
                       "/data/ginny/tcga_pancan/germline_raw_process/STAD_snpeff_annotation/snpeff_variant_anno_npc.tsv",
                     cancer_barcode = stad_barcode,
                     sample_cn_id = sample_cn_id,
}
