% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.gpd.R
\name{hist.gpd}
\alias{hist.gpd}
\title{Histogram with fit of GPD density function}
\usage{
\method{hist}{gpd}(time, censor = rep(1, times = length(time)),
  k.hat = NULL, a.hat = NULL, show.all = FALSE, ...)
}
\arguments{
\item{time}{Time to failure or termination}

\item{censor}{Observation/censor indicator (1=observed, 0=right-censored), default to be all observations}

\item{k.hat}{optional, estimate of k, estimated by mle.gpd if either k.hat or a.hat is not given}

\item{a.hat}{optional, estimate of alpha, estimated by mle.gpd if either k.hat or a.hat is not given}

\item{show.all}{(default to False) whether to draw a historam with all data points or only observations. Default to show only observed points.}

\item{...}{further arguments passed to hist()}
}
\description{
\code{hist.gpd} creates a plot of the data set's histogram and the fitted GPD density function
}
\examples{
data(melanoma, package = "boot")
status = ifelse(melanoma$status == 1, 1, 0)
hist.gpd(melanoma$time, status)

}
