% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.gpd.R
\name{mle.gpd}
\alias{mle.gpd}
\title{Parameter estimation for GPD}
\usage{
mle.gpd(time, censor = rep(1, times = length(time)))
}
\arguments{
\item{time}{Time to failure or termination}

\item{censor}{Observation/censor indicator (1=observed, 0=right-censored), default to be all observations}
}
\value{
returns a vector of size 2. The first element is the estimate of k, the second is the estimate of alpha
}
\description{
\code{mle.gpd} estimates parameters for Generalized Pareto	Distribution with right-censored data using Maximum Likelihood
}
\details{
Generalized Pareto distribution:			
CDF F(x,k,a) = 1 - (1-kx/a)^(1/k) if k!=0	or 1 - e^(-x/a) if k=0

Algorithm was described in:
Pham, M. H., Tsokos, C. P., Choi, B. (2018). Maximum Likelihood Estimation for the Generalized Pareto Distribution and Goodness-Of-Fit Test with Censored Data. Journal of Modern Applied Statistical Methods, 17(2).
}
\examples{
data(melanoma, package = "boot")
status = ifelse(melanoma$status == 1, 1, 0)
mle.gpd(melanoma$time, status)
}
