% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_association_wrap.R
\name{multivariate_cox_model_with_interaction}
\alias{multivariate_cox_model_with_interaction}
\title{Build univariate cox regression model for gene level counts with confounders adjusted (age, gender, race)}
\usage{
multivariate_cox_model_with_interaction(somatic_piu_filename,
  germline_piu_filename, somatic_bpiu_filename, germline_bpiu_filename,
  germline_npc_filename, patient_info_df, patient_outcome_df,
  piu_of_interest, row_sum_min, output_dir, output_name)
}
\arguments{
\item{somatic_piu_filename}{The name of the file having somatic PIU counts.}

\item{germline_piu_filename}{The name of the file having germline PIU counts.}

\item{somatic_bpiu_filename}{The name of the file having somatic bPIU counts.}

\item{germline_bpiu_filename}{The name of the file having germline bPIU counts.}

\item{germline_npc_filename}{The name of the file having germline npc counts.}

\item{patient_info_df}{The dataframe containing patient baseline information.}

\item{patient_outcome_df}{The dataframe containing survival information.}

\item{piu_of_interest}{Domain or PTM sites as the PIU of interest, default is "domain".}

\item{row_sum_min}{The minimum number of total occurrence for each PIU.}

\item{output_dir}{The directory you would like to have your output files in.}

\item{output_name}{The name of the output file.}
}
\description{
This function generates one file as the result, the coefficients and significance level estimated for the model.
}
\examples{
multivariate_cox_model_with_interaction(somatic_piu_filename = "/data/ginny/tcga_pancan/STAD_somatic/STAD_summarise_mutation/piu_mapping_count.tsv",
                                       germline_piu_filename = "/data/ginny/tcga_pancan/germline_raw_process/STAD_summarise_mutation/piu_mapping_count.tsv",
                                       somatic_bpiu_filename = "/data/ginny/tcga_pancan/STAD_somatic/STAD_summarise_mutation/bpiu_summarising_count.tsv",
                                       germline_bpiu_filename = "/data/ginny/tcga_pancan/germline_raw_process/STAD_summarise_mutation/bpiu_summarising_count.tsv",
                                       germline_npc_filename = "/data/ginny/tcga_pancan/germline_raw_process/STAD_summarise_mutation/npc_summarising_count.tsv",
                                       patient_info_df = patient_info,
                                       patient_outcome_df = patient_outcome,
                                       piu_of_interest = "domain",
                                       row_sum_min = 1,
                                       output_dir = "/data/ginny/tcga_pancan/STAD_somatic/cox_model/",
                                       output_name = "mutlivariate_interaction_model.tsv")
}
