% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis_wrap.R
\name{univariate_cox_model}
\alias{univariate_cox_model}
\title{This function performs survival analysis for mutation counts mapped to PIU, LU and NCU.}
\usage{
univariate_cox_model(piu_filename, lu_filename, ncu_filename, clinical_df,
  gender_as_covariate = T, race_group_min = 6, min_surv_days = 90,
  min_surv_people = 5, patient_sum_min = 3,
  mutation_type = "somatic", output_dir)
}
\arguments{
\item{piu_filename}{Filename for the PIU mapping results.}

\item{lu_filename}{Filename for the LU mapping results.}

\item{ncu_filename}{Filename for the NCU mapping results.}

\item{clinical_df}{Clinical information data frame.}

\item{gender_as_covariate}{Boolean variable indicating gender should be taken as a covariable.}

\item{race_group_min}{The minimum number of patients with the same race enough for a race group.}

\item{min_surv_days}{The minimum number of survival time (in days) a patient has to be included in analysis.}

\item{min_surv_people}{The minimum number of patients survived (or censored) in a cohort that a survival analysis should be conducted.}

\item{patient_sum_min}{The minimum number of patients having mutation mapping to the unit for the unit to be analysed.}

\item{mutation_type}{A tag indicating the type of mutation; somatic or germline.}

\item{output_dir}{The directory you would like to have your output files in.}
}
\description{
This function performs survival analysis for mutation counts mapped to PIU, LU and NCU.
}
\examples{
univariate_cox_model(piu_filename = "/Users/ginny/Google Drive/R_GPD/GPD_package_0401/example/piu_mapping_count.tsv",
                     lu_filename = "/Users/ginny/Google Drive/R_GPD/GPD_package_0401/example/lu_summarising_count.tsv",
                     ncu_filename = "/Users/ginny/Google Drive/R_GPD/GPD_package_0401/example/ncu_summarising_count.tsv",
                     clinical_df = sel_example_cdr,
                     gender_as_covariate = T,
                     race_group_min = 6,
                     min_surv_days = 90,
                     min_surv_people = 5,
                     patient_sum_min = 3,
                     mutation_type = "somatic",
                     output_dir = "/Users/ginny/Google Drive/R_GPD/GPD_package_0401/example/")
                      
}
