% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_association_wrap.R
\name{univariate_cox_model_for_gene_level_summary_count}
\alias{univariate_cox_model_for_gene_level_summary_count}
\title{Build univariate cox regression model for gene level counts with confounders adjusted (age, gender, race)}
\usage{

  univariate_cox_model_for_gene_level_summary_count(gene_level_count_filename,
  quiry_gene_id, patient_info_df, patient_outcome_df, row_sum_min = 1,
  output_dir, output_name)
}
\arguments{
\item{gene_level_count_filename}{The name of the file having gene level count of interest.}

\item{quiry_gene_id}{The set of gene ids of interest. Because we try to avoid analysing all the genes.}

\item{patient_info_df}{The dataframe containing patient baseline information.}

\item{patient_outcome_df}{The dataframe containing survival information.}

\item{row_sum_min}{The minimum number of total occurrence for each PIU.}

\item{output_dir}{The directory you would like to have your output files in.}

\item{output_name}{The name of the output file.}
}
\description{
This function generates one file as the result, the coefficients and significance level estimated for the model.
}
