% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_association_wrap.R
\name{univariate_cox_model_for_piu}
\alias{univariate_cox_model_for_piu}
\title{Build univariate cox regression model with confounders adjusted (age, gender, race)}
\usage{
univariate_cox_model_for_piu(piu_filename, patient_info_df,
  patient_outcome_df, mutation_type = "somatic",
  piu_of_interest = "domain", row_sum_min = 1, output_dir)
}
\arguments{
\item{piu_filename}{The name of the file containing PIU of interest.}

\item{patient_info_df}{The dataframe containing patient baseline information.}

\item{patient_outcome_df}{The dataframe containing survival information.}

\item{mutation_type}{Somatic or germline mutation, default is "somatic".}

\item{piu_of_interest}{Domain or PTM sites as the PIU of interest, default is "domain".}

\item{row_sum_min}{The minimum number of total occurrence for each PIU.}

\item{output_dir}{The directory you would like to have your output files in.}
}
\description{
This function generates one file as the result, the coefficients and significance level estimated for the model.
}
