% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_association_wrap.R
\name{univariate_cox_model_for_somatic_locus_piu}
\alias{univariate_cox_model_for_somatic_locus_piu}
\title{Build univariate cox regression model with confounders adjusted (age, gender, race)}
\usage{
univariate_cox_model_for_somatic_locus_piu(locus_filename, piu_filename,
  cdr_clinical, gender_as_covariate = T, race_group_min = 6,
  min_surv_days = 30, min_surv_people = 5, row_sum_min = 1,
  mutation_type = "somatic", piu_of_interest = "domain", output_dir)
}
\arguments{
\item{locus_filename}{The name of the file containing locus of interest.}

\item{piu_filename}{The name of the file containing PIU of interest.}

\item{cdr_clinical}{The cdr clinical data for the cancer type of interest.}

\item{race_group_min}{The minimum number of people to be grouped to a race group for model fitting purpose.}

\item{min_surv_days}{The minimum number of days only above which the associated patient is selected for model building.}

\item{min_surv_people}{The minimum number of people on both alive and deceased sides.}

\item{row_sum_min}{The minimum number of total occurrence for each PIU.}

\item{mutation_type}{Somatic or germline mutation, default is "somatic".}

\item{piu_of_interest}{Domain or PTM sites as the PIU of interest, default is "domain".}

\item{output_dir}{The directory you would like to have your output files in.}
}
\description{
This function generates one file as the result, the coefficients and significance level estimated for the model.
}
\examples{
univariate_cox_model_for_somatic_locus_piu(locus_filename = "/data/ginny/tcga_pancan/STAD_somatic/STAD_summarise_mutation/stad_mc3_count_matrix.tsv",
                                          piu_filename = "/data/ginny/tcga_pancan/STAD_somatic/STAD_summarise_mutation/piu_mapping_count.tsv",
                                          cdr_clinical = stad_cdr,
                                          race_group_min = 6,
                                          min_surv_days = 30,
                                          min_surv_people = 5,
                                          row_sum_min = 1,
                                          mutation_type = "somatic",
                                          piu_of_interest = "domain",
                                          output_dir = "/data/ginny/tcga_pancan/STAD_somatic/cox_model/")
}
