% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_extraction_mapping_wrap.R
\name{extraction_annotation_pos}
\alias{extraction_annotation_pos}
\title{This function generates two files as the final products. First, mutations which result in protein-consequence with
effect on proteins annotated; Second, mutations without protein-consequence.}
\usage{
extraction_annotation_pos(mutation_df, cancer_type, cancer_barcode,
  output_dir)
}
\arguments{
\item{mutation_df}{Data frame of mc3 somatic mutations.}

\item{cancer_type}{An identifier of cancer type, e.g. ACC,STAD,BRCA.}

\item{output_dir}{The directory you would like to have your output files in.}

\item{cancer_barcod}{A list of barcodes user wants to process, matching the last column of mutation file.}
}
\description{
This function generates two files as the final products. First, mutations which result in protein-consequence with
effect on proteins annotated; Second, mutations without protein-consequence.
}
\examples{
extraction_annotation_pos(mutation_df = sel_example_mutation,
                                 cancer_type = "example",
                                 cancer_barcode = sel_example_barcode,
                                 output_dir = "/Users/ginny/Google Drive/R_GPD/GPD_package_0401/example/")
}
