% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_extraction_mapping_wrap.R
\name{piu_mapping}
\alias{piu_mapping}
\title{This function generates three files as the final products. First, each piu mapped, second, lu summarised per gene, third, 
npc summarised per gene.}
\usage{
piu_mapping(piu_df, pc_data_name, npc_data_name, cancer_barcode,
  output_dir)
}
\arguments{
\item{pc_data_name}{the file listing mutations that result in protein consequence. A file generated by function germline_extraction_annotation_pos}

\item{npc_data_name}{the file listing mutations that result in non-protein consequence. A file generated by function germline_extraction_annotation_pos}

\item{cancer_barcode}{Patient barcodes for this cancer type cohort.}

\item{output_dir}{The directory you would like to have your output files in.}

\item{piu_filename}{The file provided by the package where the piu information is recorded.}
}
\description{
This function generates three files as the final products. First, each piu mapped, second, lu summarised per gene, third, 
npc summarised per gene.
}
\examples{
piu_mapping (piu_df = ptm_pfam_combine,
                     pc_data_name  =  "/Users/ginny/Google Drive/R_GPD/GPD_package_0401/example/example_mutation_pc_pos.tsv",
                     npc_data_name = "/Users/ginny/Google Drive/R_GPD/GPD_package_0401/example/example_mutation_npc.tsv",
                     cancer_barcode = sel_example_barcode,
}
